# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Fluxend is backend as-a-service platform that allows you to build, deploy, and scale applications without managing infrastructure.",
        "title": "Fluxend API",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/fluxend/fluxend",
            "email": "hello@fluxend.app"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/functions/{schema}": {
            "get": {
                "description": "Retrieve a list of all functions for the specified schema",
                "tags": [
                    "Functions"
                ],
                "summary": "List functions",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Project UUID",
                        "name": "X-Project",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Project UUID",
                        "name": "projectUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Schema to search under",
                        "name": "schema",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of functions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "allOf": [
                                            {
                                                "$ref": "#/components/schemas/response.Response"
                                            },
                                            {
                                                "type": "object",
                                                "properties": {
                                                    "content": {
                                                        "type": "array",
                                                        "items": {
                                                            "$ref": "#/components/schemas/database.FunctionResponse"
                                                        }
                                                    }
                                                }
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "description": "Add a new function for specific schema",
                "tags": [
                    "Functions"
                ],
                "summary": "Create function",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Project UUID",
                        "name": "X-Project",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/database.CreateFunctionRequest"
                            }
                        }
                    },
                    "description": "Function details",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Function created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {
                                                    "$ref": "#/components/schemas/database.FunctionResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/functions/{schema}/{functionName}": {
            "get": {
                "description": "Get details of a specific function",
                "tags": [
                    "Functions"
                ],
                "summary": "Retrieve function",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Project UUID",
                        "name": "X-Project",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Schema name",
                        "name": "schema",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Function name",
                        "name": "functionName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Function details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {
                                                    "$ref": "#/components/schemas/database.FunctionResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "description": "Remove a function from the schema",
                "tags": [
                    "Functions"
                ],
                "summary": "Delete function",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Project UUID",
                        "name": "X-Project",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Project UUID",
                        "name": "projectUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Schema name",
                        "name": "schema",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Function name",
                        "name": "functionName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Form deleted"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//api.fluxend.app"
        }
    ],
    "components": {
...
        "schemas": {
...
            "database.CreateFunctionRequest": {
                "type": "object",
                "properties": {
                    "definition": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "parameters": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/database.functionParameter"
                        }
                    },
                    "projectUUID": {
                        "type": "string"
                    },
                    "return_type": {
                        "type": "string"
                    }
                }
            },
...
            "database.FunctionResponse": {
                "type": "object",
                "properties": {
                    "dataType": {
                        "type": "string"
                    },
                    "definition": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
...
            "database.functionParameter": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
...
            "response.Response": {
                "type": "object",
                "properties": {
                    "content": {},
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}